test update and small testing guide
