+++ test/localStorage/localStorageTestData.js
@@ -1,5 +1,3 @@
-const { addResult } = require('../src/app/logic/localStorage');
-
 const testGameInfos = [
   {
     name: 'Albus Dumbledore',
@@ -53,22 +51,4 @@ const testGameInfos = [
   },
 ];
 
-function testLocalStorage(rankings) {
-  const button = document.createElement('button');
-  button.innerText = 'TEST - Save gameInfos in localStorage';
-  button.setAttribute('id', 'localstorage');
-  document.querySelector('#hpquiz-app').append(button);
-
-  const buttonLS = document.getElementById('localstorage');
-  buttonLS.addEventListener('click', (event) => {
-    event.preventDefault();
-    buttonLS.innerText = 'TEST - Now check the localStorage';
-    for (const info of testGameInfos) {
-      addResult(rankings, info, 3);
-    }
-  });
-}
-
-module.exports = {
-  testLocalStorage,
-};
+export default testGameInfos;
